+++ renderlib/ImageXYZC.cpp
@@ -26,7 +26,7 @@ ImageXYZC::ImageXYZC(uint32_t x,
   , m_scaleY(sy)
   , m_scaleZ(sz)
   , m_spatialUnits(spatialUnits)
-  , m_flipped(1.0, 1.0, 1.0)
+  , m_flipped(1, 1, 1)
 {
   for (uint32_t i = 0; i < m_c; ++i) {
     m_channels.push_back(new Channelu16(x, y, z, reinterpret_cast<uint16_t*>(ptr(i))));
@@ -86,12 +86,12 @@ ImageXYZC::setPhysicalSize(float x, float y, float z)
 }
 
 void
-ImageXYZC::setVolumeAxesFlipped(float x, float y, float z)
+ImageXYZC::setVolumeAxesFlipped(int x, int y, int z)
 {
-  m_flipped = glm::vec3(x < 0 ? -1 : 1, y < 0 ? -1 : 1, z < 0 ? -1 : 1);
+  m_flipped = glm::ivec3(x < 0 ? -1 : 1, y < 0 ? -1 : 1, z < 0 ? -1 : 1);
 }
 
-glm::vec3
+glm::ivec3
 ImageXYZC::getVolumeAxesFlipped() const
 {
   return m_flipped;
